automate/requirements.txt
Django==5.2
django-import-export==4.3.7
Faker==37.1.0
openpyxl==3.1.5
paho-mqtt==2.1.0
pandas==2.2.3
requests==2.32.3